feat(album): refactor album module by creating separate types and utility functions for better organization and maintainability
